@@ -1,65 +1,102 @@
 name: allo
 description: Chat app
 publish_to: 'none'
-version: 1.3.3+210
+version: 1.3.5+330
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  animations: ^2.0.2
-  awesome_notifications: ^0.6.21
-  cached_network_image: ^3.2.0
-  cloud_firestore: ^3.1.6
-  device_info_plus: ^3.2.2
-  firebase_auth: ^3.3.5
-  firebase_core: ^1.12.0
-  firebase_image: ^1.1.1
-  firebase_messaging: ^11.2.6
-  firebase_remote_config: ^2.0.0
-  firebase_storage: ^10.2.6
+  animations: ^2.0.5
+  awesome_notifications: ^0.7.1
+  animated_progress:
+    git: https://github.com/iamcosmin/animated_progress
+  cached_network_image: ^3.2.2
+  cloud_firestore: ^3.5.0
+  device_info_plus: ^5.0.5
+  dynamic_color: ^1.5.3
+  firebase_auth: ^3.11.0
+  firebase_core: ^1.24.0
+  firebase_messaging: ^13.0.0
+  firebase_remote_config: ^2.0.19
+  firebase_storage: ^10.3.10
+  firebase_image:
+    path: ./packages/firebase_image
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.18.0
-  flutter_linkify: ^5.0.2
   flutter_localizations:
     sdk: flutter
-  flutter_platform_widgets: ^1.20.0
-  flutterfire_ui: ^0.3.2
-  hooks_riverpod: ^1.0.1
-  image_cropper: ^1.4.1
-  image_picker: ^0.8.4+6
-  in_app_update: ^2.0.0
+  flutter_hooks: ^0.18.5+1
+  flutter_linkify: ^5.0.2
+  hooks_riverpod: ^2.0.2
+  http: ^0.13.4
+  image_cropper: ^3.0.0
+  image_picker: ^0.8.5+3
+  # in_app_update: ^2.0.0 // Some bugs do not permit for this to run. We will remove it temporarily as it is not used either.
   intl: ^0.17.0
-  package_info_plus: ^1.4.0
-  shared_preferences: ^2.0.13
-  url_launcher: ^6.0.18
+  package_info_plus: ^1.4.2
+  shared_preferences: ^2.0.15
+  url_launcher: ^6.1.6
+  material_color_utilities: ^0.2.0
+  shimmer: ^2.0.0
+  sliver_tools: ^0.2.8
+  flutter_staggered_animations: ^1.1.1
+  snap_scroll_physics: ^0.0.1+3
+  just_audio: ^0.9.29
+  go_router: ^5.1.1
+  stack_trace: ^1.10.0
+  # logger: ^1.1.0
+  # logger_flutter: ^0.7.1
 
 dev_dependencies:
-  flutter_lints: ^1.0.4
+  dart_code_metrics: ^4.21.2
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  pedantic: ^1.11.1
 
 flutter:
-  uses-material-design: true
   generate: true
+  uses-material-design: true
+
+  assets:
+   - assets/images/
+   - assets/audio/
+
   fonts:
-    - family: GS-Text
+    - family: Jakarta
       fonts:
-        - asset: fonts/GS-Text-Regular.ttf
-        - asset: fonts/GS-Text-Bold.ttf
-          weight: 700
-        - asset: fonts/GS-Text-Bold-Italic.ttf
-          weight: 700
+        - asset: assets/fonts/Jakarta-ExtraLight.ttf
+          weight: 200
+        - asset: assets/fonts/Jakarta-ExtraLightItalic.ttf
+          weight: 200
           style: italic
-        - asset: fonts/GS-Text-Medium.ttf
+        - asset: assets/fonts/Jakarta-Light.ttf
+          weight: 300
+        - asset: assets/fonts/Jakarta-LightItalic.ttf
+          weight: 300
+          style: italic
+        - asset: assets/fonts/Jakarta-Regular.ttf
+          weight: 400
+        - asset: assets/fonts/Jakarta-Italic.ttf
+          weight: 400
+          style: italic
+        - asset: assets/fonts/Jakarta-Medium.ttf
           weight: 500
-        - asset: fonts/GS-Text-Medium-Italic.ttf
+        - asset: assets/fonts/Jakarta-MediumItalic.ttf
           weight: 500
           style: italic
-        - asset: fonts/GS-Text-Italic.ttf
+        - asset: assets/fonts/Jakarta-SemiBold.ttf
+          weight: 600
+        - asset: assets/fonts/Jakarta-SemiBoldItalic.ttf
+          weight: 600
           style: italic
-flutter_intl:
-  enabled: true
-  main_locale: en
+        - asset: assets/fonts/Jakarta-Bold.ttf
+          weight: 700
+        - asset: assets/fonts/Jakarta-BoldItalic.ttf
+          weight: 700
+          style: italic
+        - asset: assets/fonts/Jakarta-ExtraBold.ttf
+          weight: 800
+        - asset: assets/fonts/Jakarta-ExtraBoldItalic.ttf
+          weight: 800
+          style: italic
