@@ -1,3 +1,5 @@
+// ignore_for_file: prefer_null_aware_method_calls
+
 import 'package:flutter/material.dart';
 
 class PinchZoom extends StatefulWidget {
@@ -23,20 +25,20 @@ class PinchZoom extends StatefulWidget {
   /// * [onZoomStart] called when the widget goes to its zoomed state.
   ///
   /// * [onZoomEnd] called when the widget is back to its idle state.
-  const PinchZoom(
-      {Key? key,
-      required this.child,
-      this.resetDuration = const Duration(milliseconds: 100),
-      // This default maxScale value is eyeballed as reasonable limit for common
-      // use cases.
-      this.maxScale = 3.0,
-      this.zoomEnabled = true,
-      this.onZoomStart,
-      this.onZoomEnd})
-      : super(key: key);
+  const PinchZoom({
+    required this.child,
+    super.key,
+    this.resetDuration = const Duration(milliseconds: 100),
+    // This default maxScale value is eyeballed as reasonable limit for common
+    // use cases.
+    this.maxScale = 3.0,
+    this.zoomEnabled = true,
+    this.onZoomStart,
+    this.onZoomEnd,
+  });
 
   @override
-  _PinchZoomState createState() => _PinchZoomState();
+  State<PinchZoom> createState() => _PinchZoomState();
 }
 
 class _PinchZoomState extends State<PinchZoom>
@@ -53,7 +55,6 @@ class _PinchZoomState extends State<PinchZoom>
       height: double.maxFinite,
       width: double.maxFinite,
       child: InteractiveViewer(
-        child: widget.child,
         scaleEnabled: widget.zoomEnabled,
         maxScale: widget.maxScale,
         panEnabled: false,
@@ -70,6 +71,7 @@ class _PinchZoomState extends State<PinchZoom>
             : null,
         onInteractionEnd: (_) => _animateResetInitialize(),
         transformationController: _transformationController,
+        child: widget.child,
       ),
     );
   }
