@@ -1,59 +1,54 @@
-import 'package:allo/generated/l10n.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
+import '../../../../components/slivers/sliver_scaffold.dart';
+import '../../../../components/slivers/top_app_bar.dart';
+
 class AccountInfo extends HookWidget {
-  const AccountInfo({Key? key}) : super(key: key);
+  const AccountInfo({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final username = useState('');
-    final locales = S.of(context);
-    useEffect(() {
-      Future.microtask(() async {
-        username.value = await Core.auth.user.username;
-      });
-      return;
-    }, const []);
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(locales.internalAccountInfo),
+    final username = useFuture(Core.auth.user.getUsername());
+    return SScaffold(
+      topAppBar: LargeTopAppBar(
+        title: Text(context.loc.internalAccountInfo),
       ),
-      body: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            SelectableText(
-              '${locales.name}: ' + Core.auth.user.name,
-              style: const TextStyle(fontSize: 17),
-            ),
-            const Padding(padding: EdgeInsets.only(top: 20)),
-            SelectableText(
-              '${locales.initials}: ' + Core.auth.user.nameInitials,
-              style: const TextStyle(fontSize: 17),
-            ),
-            const Padding(padding: EdgeInsets.only(top: 20)),
-            SelectableText(
-              '${locales.uid}: ' + Core.auth.user.uid,
-              style: const TextStyle(fontSize: 17),
-            ),
-            const Padding(padding: EdgeInsets.only(top: 20)),
-            SelectableText(
-              '${locales.username}: ' + username.value,
-              style: const TextStyle(fontSize: 17),
-            ),
-            const Padding(padding: EdgeInsets.only(top: 20)),
-            SelectableText(
-              '${locales.profilePicture} ' +
-                  (Core.auth.user.profilePicture ?? locales.noProfilePicture),
-              style: const TextStyle(fontSize: 17),
-            ),
-            const Padding(padding: EdgeInsets.only(top: 20)),
-          ],
-        ),
-      ),
+      slivers: [
+        SliverPadding(
+          padding: const EdgeInsets.all(20),
+          sliver: SliverList(
+            delegate: SliverChildListDelegate.fixed([
+              SelectableText(
+                '${context.loc.name}: ${Core.auth.user.name}',
+                style: const TextStyle(fontSize: 17),
+              ),
+              const Padding(padding: EdgeInsets.only(top: 20)),
+              SelectableText(
+                '${context.loc.initials}: ${Core.auth.user.nameInitials}',
+                style: const TextStyle(fontSize: 17),
+              ),
+              const Padding(padding: EdgeInsets.only(top: 20)),
+              SelectableText(
+                '${context.loc.uid}: ${Core.auth.user.userId}',
+                style: const TextStyle(fontSize: 17),
+              ),
+              const Padding(padding: EdgeInsets.only(top: 20)),
+              SelectableText(
+                '${context.loc.username}: ${username.data}',
+                style: const TextStyle(fontSize: 17),
+              ),
+              const Padding(padding: EdgeInsets.only(top: 20)),
+              SelectableText(
+                '${context.loc.profilePicture} ${Core.auth.user.profilePictureUrl ?? context.loc.noProfilePicture}',
+                style: const TextStyle(fontSize: 17),
+              ),
+              const Padding(padding: EdgeInsets.only(top: 20)),
+            ]),
+          ),
+        )
+      ],
     );
   }
 }
