@@ -1,108 +1,103 @@
-import 'package:allo/components/oobe_page.dart';
-import 'package:allo/generated/l10n.dart';
-import 'package:allo/interface/login/new/setup_username.dart';
+import 'package:allo/components/setup_page.dart';
+import 'package:allo/logic/backend/setup/login.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:go_router/go_router.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class SetupName extends HookWidget {
-  const SetupName(this.email, {Key? key}) : super(key: key);
-  final String email;
+final _nameReg = RegExp(r'^[a-zA-Z]+$');
+
+class SetupName extends HookConsumerWidget {
+  const SetupName({super.key});
   @override
-  Widget build(BuildContext context) {
-    final locales = S.of(context);
-    final firstFieldError = useState('');
-    final secondFieldError = useState('');
+  Widget build(BuildContext context, ref) {
+    final firstFieldError = useState<String?>(null);
+    final secondFieldError = useState<String?>(null);
     final firstNameController = useTextEditingController();
     final secondNameController = useTextEditingController();
-    final nameReg = RegExp(r'^[a-zA-Z]+$');
-    return SetupPage(
-      header: [
-        Text(
-          locales.setupNameScreenTitle,
-          style: const TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-          textAlign: TextAlign.left,
-        ),
-        const Padding(
-          padding: EdgeInsets.only(top: 10),
-        ),
-        Text(
-          locales.setupNameScreenDescription,
-          style: const TextStyle(fontSize: 17, color: Colors.grey),
-          textAlign: TextAlign.left,
-        ),
-      ],
-      body: [
-        Padding(
-          padding: const EdgeInsets.only(left: 20, right: 20),
-          child: Column(
-            children: [
-              TextFormField(
-                decoration: InputDecoration(
-                  contentPadding: const EdgeInsets.all(10),
-                  errorText: firstFieldError.value == ''
-                      ? null
-                      : firstFieldError.value,
-                  errorStyle: const TextStyle(fontSize: 14),
-                  labelText: locales.firstName,
-                  border: const OutlineInputBorder(),
-                ),
-                controller: firstNameController,
-              ),
-              const Padding(padding: EdgeInsets.only(bottom: 10)),
-              TextFormField(
-                decoration: InputDecoration(
-                  contentPadding: const EdgeInsets.all(10),
-                  errorText: secondFieldError.value == ''
-                      ? null
-                      : secondFieldError.value,
-                  errorStyle: const TextStyle(fontSize: 14),
-                  labelText: locales.lastName,
-                  border: const OutlineInputBorder(),
-                ),
-                controller: secondNameController,
-              ),
-            ],
-          ),
-        ),
-      ],
-      onButtonPress: () {
-        FocusScope.of(context).unfocus();
-        firstFieldError.value = '';
-        secondFieldError.value = '';
-        if (firstNameController.text != '') {
-          if (nameReg.hasMatch(firstNameController.text)) {
-            if (secondNameController.text != '') {
-              if (nameReg.hasMatch(secondNameController.text)) {
-                Core.navigation.push(
-                  context: context,
-                  route: SetupUsername(
-                    displayName: firstNameController.text +
-                        ' ' +
-                        secondNameController.text,
-                    email: email,
-                  ),
-                );
-              } else {
-                secondFieldError.value = locales.specialCharactersNotAllowed;
-              }
+    final focusController = useFocusNode();
+
+    void validateName(String name, ValueNotifier<String?> error) {
+      if (name.isNotEmpty && !_nameReg.hasMatch(name)) {
+        error.value = context.loc.specialCharactersNotAllowed;
+      } else {
+        error.value = null;
+      }
+    }
+
+    void onSubmit() {
+      FocusScope.of(context).unfocus();
+      if (firstNameController.text.isNotEmpty) {
+        if (_nameReg.hasMatch(firstNameController.text)) {
+          if (secondNameController.text.isNotEmpty) {
+            if (_nameReg.hasMatch(secondNameController.text)) {
+              ref.read(signupState.notifier).addName(
+                    '${firstNameController.text} ${secondNameController.text}',
+                  );
+              context.go('/start/signup/username');
             } else {
-              Core.navigation.push(
-                context: context,
-                route: SetupUsername(
-                  displayName: firstNameController.text,
-                  email: email,
-                ),
-              );
+              secondFieldError.value = context.loc.specialCharactersNotAllowed;
             }
           } else {
-            firstFieldError.value = locales.specialCharactersNotAllowed;
+            ref.read(signupState.notifier).addName(firstNameController.text);
+            context.go('/start/signup/username');
           }
         } else {
-          firstFieldError.value = locales.errorFieldEmpty;
+          firstFieldError.value = context.loc.specialCharactersNotAllowed;
         }
-      },
-      isAsync: true,
+      } else {
+        firstFieldError.value = context.loc.errorFieldEmpty;
+      }
+    }
+
+    return SetupPage(
+      icon: Icons.person,
+      title: Text(context.loc.setupNameScreenTitle),
+      subtitle: Text(context.loc.setupNameScreenDescription),
+      body: [
+        Column(
+          children: [
+            TextFormField(
+              decoration: InputDecoration(
+                contentPadding: const EdgeInsets.all(10),
+                errorText: firstFieldError.value,
+                errorStyle: const TextStyle(fontSize: 14),
+                labelText: context.loc.firstName,
+                border: const OutlineInputBorder(),
+              ),
+              autofocus: true,
+              onChanged: (_) {
+                validateName(_, firstFieldError);
+              },
+              onFieldSubmitted: (_) {
+                validateName(_, firstFieldError);
+                if (firstFieldError.value == null) {
+                  focusController.requestFocus();
+                }
+              },
+              controller: firstNameController,
+            ),
+            const Padding(padding: EdgeInsets.only(bottom: 10)),
+            TextFormField(
+              focusNode: focusController,
+              decoration: InputDecoration(
+                contentPadding: const EdgeInsets.all(10),
+                errorText: secondFieldError.value,
+                errorStyle: const TextStyle(fontSize: 14),
+                labelText: context.loc.lastName,
+                border: const OutlineInputBorder(),
+              ),
+              onChanged: (_) {
+                validateName(_, secondFieldError);
+              },
+              onFieldSubmitted: (_) => onSubmit(),
+              controller: secondNameController,
+            ),
+          ],
+        ),
+      ],
+      action: onSubmit,
     );
   }
 }
