@@ -1,60 +1,56 @@
-import 'package:allo/generated/l10n.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/interface/login/new/setup_pfp.dart';
 import 'package:allo/logic/core.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:allo/components/oobe_page.dart';
+
+const _kDebugBypassVerification = kDebugMode ? true : false;
 
 class SetupVerification extends HookWidget {
-  const SetupVerification({Key? key}) : super(key: key);
+  const SetupVerification({
+    super.key,
+  });
   @override
   Widget build(BuildContext context) {
-    final locales = S.of(context);
-    final error = useState('');
-
     useEffect(() {
-      Future.microtask(() async =>
-          await FirebaseAuth.instance.currentUser!.sendEmailVerification());
+      Future.microtask(
+        () async =>
+            await FirebaseAuth.instance.currentUser!.sendEmailVerification(),
+      );
       return;
     });
     return SetupPage(
-        header: [
-          Text(
-            locales.setupVerificationScreenTitle,
-            style: const TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-            textAlign: TextAlign.left,
-          ),
-          Padding(
-            padding: const EdgeInsets.only(right: 10, top: 10),
-            child: Text(
-              locales.setupVerificationScreenDescription,
-              style: const TextStyle(fontSize: 17, color: Colors.grey),
-              textAlign: TextAlign.left,
+      icon: Icons.email,
+      title: Text(context.loc.setupVerificationScreenTitle),
+      subtitle: Text(context.loc.setupVerificationScreenDescription),
+      action: () async {
+        await FirebaseAuth.instance.currentUser?.reload();
+        final verified = FirebaseAuth.instance.currentUser!.emailVerified;
+        if (verified || _kDebugBypassVerification) {
+          /* TODO: Theoretically, this will not be executed, as the app will automatically redirect the user to the chats
+          screen if the app detects that the email has been verified.*/
+
+          // Awaiting test result for deleting supplimentary configuration.
+          // Another option would be making some banners into the chats screen recommending the user to add a profile picture, etc.
+          Navigation.forward(const SetupProfilePicture());
+        } else {
+          await showDialog(
+            context: context,
+            builder: (context) => AlertDialog(
+              title: Text(context.loc.error),
+              content: Text(context.loc.errorVerificationLinkNotAccessed),
+              actions: [
+                TextButton(
+                  onPressed: () => Navigator.pop(context),
+                  child: const Text('OK'),
+                ),
+              ],
             ),
-          ),
-          Padding(
-            padding: const EdgeInsets.only(right: 10, top: 10),
-            child: Text(
-              error.value,
-              style: const TextStyle(fontSize: 17, color: Colors.red),
-              textAlign: TextAlign.left,
-            ),
-          )
-        ],
-        body: const [],
-        onButtonPress: () async {
-          await FirebaseAuth.instance.currentUser?.reload();
-          final verified = FirebaseAuth.instance.currentUser!.emailVerified;
-          if (verified) {
-            await Core.navigation.push(
-              context: context,
-              route: const SetupProfilePicture(),
-            );
-          } else {
-            error.value = locales.errorVerificationLinkNotAccessed;
-          }
-        },
-        isAsync: true);
+          );
+        }
+      },
+    );
   }
 }
