@@ -1,139 +1,165 @@
-import 'dart:io';
-
-import 'package:allo/generated/l10n.dart';
-import 'package:allo/interface/login/main_setup.dart';
+import 'package:allo/logic/backend/firebase_options.dart';
+import 'package:allo/logic/client/navigation/routing.dart';
+import 'package:allo/logic/client/preferences/manager.dart';
+import 'package:allo/logic/client/preferences/preferences.dart';
+import 'package:allo/logic/client/theme/theme.dart';
 import 'package:allo/logic/core.dart';
-import 'package:allo/logic/preferences.dart';
-import 'package:allo/logic/theme.dart';
-import 'package:allo/logic/themes.dart';
-import 'package:allo/logic/types.dart';
-import 'package:awesome_notifications/awesome_notifications.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:animated_progress/animated_progress.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'interface/home/chat/chat.dart';
-import 'interface/home/tabbed_navigator.dart';
-import 'logic/chat/chat.dart';
-import 'logic/notifications.dart';
+import 'package:stack_trace/stack_trace.dart' as stack_trace;
 
-final navigatorKey = GlobalKey<NavigatorState>();
+import 'components/space.dart';
 
-void main() async {
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  if (kIsWeb) {
-    await Firebase.initializeApp(
-      options: const FirebaseOptions(
-        apiKey: "AIzaSyAyLt2_FAHc0I2c1iBLH_MxWzo2kllSvA8",
-        authDomain: "allo-ms.firebaseapp.com",
-        projectId: "allo-ms",
-        storageBucket: "allo-ms.appspot.com",
-        messagingSenderId: "1049075385887",
-        appId: "1:1049075385887:web:89f4887e574f8b93f2372a",
-        measurementId: "G-N5D9CRB413",
-      ),
-    );
-  } else if (Platform.isAndroid) {
-    // Todo: Migrate to dart only initialisation for android.
-    await Firebase.initializeApp();
-  }
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+  await Notifications.setupNotifications();
+  await FirebaseRemoteConfig.instance.fetchAndActivate();
+  // Temporary, once framework fixes SystemUiMode.edgeToEdge, we can return to it.
+  SystemChrome.setSystemUIOverlayStyle(
+    const SystemUiOverlayStyle(
+      statusBarColor: Colors.transparent,
+    ),
+  );
 
-  if (!kIsWeb) {
-    await Core.notifications.setupNotifications();
-    FirebaseMessaging.onBackgroundMessage(onBackgroundMessage);
-  }
-  final _kSharedPreferences = await SharedPreferences.getInstance();
-  // await FirebaseMessaging.instance.requestPermission(
-  //   alert: true,
-  //   announcement: false,
-  //   badge: true,
-  //   carPlay: false,
-  //   criticalAlert: false,
-  //   provisional: false,
-  //   sound: true,
-  // );
-  // await FirebaseMessaging.instance.getToken(
-  //   vapidKey:
-  //       'BAx5uT7szCuYzwq9fLUNwS9-OF-GwOa4eGAb5J3jfl2d3e3L2b354oRm89KQ6sUbiEsK5YLPJoOs0n25ibcGbO8',
-  // );
   runApp(
     ProviderScope(
-      overrides: [
-        sharedPreferencesProvider.overrideWithValue(_kSharedPreferences),
-      ],
-      child: const InnerApp(),
+      overrides: await Keys.getOverrides(),
+      child: const App(),
     ),
   );
 }
 
-class InnerApp extends HookConsumerWidget {
-  const InnerApp({Key? key}) : super(key: key);
+class App extends ConsumerWidget {
+  const App({super.key});
+
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final darkState = ref.watch(darkMode);
-    const _scrollBehavior = MaterialScrollBehavior(
-        androidOverscrollIndicator: AndroidOverscrollIndicator.stretch);
-    useEffect(() {
-      FirebaseRemoteConfig.instance.fetchAndActivate();
-      return;
-    }, const []);
-    useNotificationListener(context);
-    return MaterialApp(
+  Widget build(context, ref) {
+    Notifications.ensureListenersActive();
+    final darkState = useSetting(ref, darkMode);
+    ref.watch(Core.auth.stateProvider);
+    FlutterError.demangleStackTrace = (stack) {
+      if (stack is stack_trace.Trace) return stack.vmTrace;
+      if (stack is stack_trace.Chain) return stack.toTrace().vmTrace;
+      return stack;
+    };
+    return MaterialApp.router(
       title: 'Allo',
       debugShowCheckedModeBanner: false,
-      navigatorKey: navigatorKey,
-      scrollBehavior: _scrollBehavior,
-      themeMode: darkState ? ThemeMode.dark : ThemeMode.light,
-      theme: theme(Brightness.light, ref, context),
-      darkTheme: theme(Brightness.dark, ref, context),
-      localizationsDelegates: const [
-        S.delegate,
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        GlobalCupertinoLocalizations.delegate,
-      ],
-      supportedLocales: S.delegate.supportedLocales,
-      home: CupertinoTheme(
-        data: CupertinoThemeData(
-            brightness: darkState ? Brightness.dark : Brightness.light,
-            primaryColor: theme(darkState ? Brightness.dark : Brightness.light,
-                    ref, context)
-                .colorScheme
-                .primary,
-            scaffoldBackgroundColor: theme(
-                    darkState ? Brightness.dark : Brightness.light,
-                    ref,
-                    context)
-                .colorScheme
-                .surface),
-        child: StreamBuilder(
-          stream: FirebaseAuth.instance.authStateChanges(),
-          builder: (context, snap) {
-            if (snap.hasData) {
-              return TabbedNavigator();
-            } else if (snap.connectionState == ConnectionState.waiting) {
-              return const Scaffold(
-                body: Center(
-                  child: SizedBox(
-                    height: 60,
-                    width: 60,
-                    child: CircularProgressIndicator(),
-                  ),
-                ),
-              );
-            } else {
-              return const Setup();
-            }
+      scaffoldMessengerKey: Keys.scaffoldMessengerKey,
+      color: kDefaultBrandingColor,
+      themeMode:
+          ThemeMode.values.firstWhere((_) => _.toString() == darkState.setting),
+      theme: theme(Brightness.light, ref),
+      darkTheme: theme(Brightness.dark, ref),
+      localizationsDelegates: AppLocalizations.localizationsDelegates,
+      supportedLocales: AppLocalizations.supportedLocales,
+
+      /// ROUTER
+      routerConfig: ref.read(routing),
+      builder: (context, child) {
+        return ref.watch(Core.auth.stateProvider).when(
+          data: (data) {
+            return child!;
           },
-        ),
-      ),
+          error: (error, stack) {
+            return Center(
+              child: Text(error.toString()),
+            );
+          },
+          loading: () {
+            return Scaffold(
+              body: Center(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      context.loc.startingApp,
+                      style: context.textTheme.titleMedium!.copyWith(
+                        color: context.colorScheme.onSurface,
+                      ),
+                    ),
+                    const Space(0.5),
+                    const ProgressBar(),
+                  ],
+                ),
+              ),
+            );
+          },
+        );
+      },
     );
   }
 }
+
+// TODO: Prepare for removal.
+// class InnerApp extends HookConsumerWidget {
+//   const InnerApp({super.key});
+//   @override
+//   Widget build(BuildContext context, WidgetRef ref) {
+//     final darkState = useSetting(ref, darkMode);
+//     final authState = ref.watch(Core.auth.stateProvider);
+//     if (!kIsWeb) {
+//       useEffect(
+//         () {
+//           Notifications.ensureListenersActive();
+//           return;
+//         },
+//         const [],
+//       );
+//     }
+//     return MaterialApp(
+//       title: 'Allo',
+//       debugShowCheckedModeBanner: false,
+//       navigatorKey: rootNavigatorKey,
+//       scaffoldMessengerKey: Keys.scaffoldMessengerKey,
+//       themeMode: ThemeMode.values.firstWhere(
+//         (element) => darkState.setting == element.toString(),
+//       ),
+//       theme: theme(Brightness.light, ref),
+//       darkTheme: theme(Brightness.dark, ref),
+//       localizationsDelegates: AppLocalizations.localizationsDelegates,
+//       supportedLocales: AppLocalizations.supportedLocales,
+//       home: authState.when(
+//         data: (data) {
+//           return PageTransitionSwitcher(
+//             transitionBuilder: (child, primaryAnimation, secondaryAnimation) {
+//               return child;
+//             },
+//             child: Builder(
+//               builder: (context) {
+//                 switch (data) {
+//                   case AuthState.emailNotVerified:
+//                     return const EmailNotVerifiedPage();
+//                   case AuthState.signedOut:
+//                     return const IntroPage();
+//                   case AuthState.signedIn:
+//                     return const TabbedNavigator(Home());
+//                 }
+//               },
+//             ),
+//           );
+//         },
+//         error: (_, __) => const Center(
+//           child: Text('Error!'),
+//         ),
+//         loading: () {
+//           return const Scaffold(
+//             body: Center(
+//               child: SizedBox(
+//                 height: 60,
+//                 width: 60,
+//                 child: ProgressRing(),
+//               ),
+//             ),
+//           );
+//         },
+//       ),
+//     );
+//   }
+// }
