@@ -1,120 +1,111 @@
-import 'package:allo/components/oobe_page.dart';
-import 'package:allo/generated/l10n.dart';
+import 'package:allo/components/setup_page.dart';
+import 'package:allo/logic/backend/setup/login.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class SetupPassword extends HookWidget {
-  const SetupPassword(
-      {required this.displayName,
-      required this.username,
-      required this.email,
-      Key? key})
-      : super(key: key);
-  final String displayName;
-  final String username;
-  final String email;
+class SetupPassword extends HookConsumerWidget {
+  const SetupPassword({
+    super.key,
+  });
   @override
-  Widget build(BuildContext context) {
-    final error = useState('');
-    final locales = S.of(context);
+  Widget build(context, ref) {
+    final error = useState<String?>(null);
     final obscure = useState(true);
     final passController = useTextEditingController();
     final confirmPassController = useTextEditingController();
-
     return SetupPage(
-      header: [
-        Text(
-          locales.setupPasswordScreenTitle,
-          style: const TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-          textAlign: TextAlign.left,
-        ),
-        const Padding(
-          padding: EdgeInsets.only(top: 10),
-        ),
-        Text(
-          locales.setupPasswordScreenDescription,
-          style: const TextStyle(fontSize: 17, color: Colors.grey),
-          textAlign: TextAlign.left,
-        ),
-      ],
+      icon: Icons.password,
+      title: Text(context.loc.setupPasswordScreenTitle),
+      subtitle: Text(context.loc.setupPasswordScreenDescription),
       body: [
-        Padding(
-          padding: const EdgeInsets.only(left: 20, right: 20),
-          child: Column(
-            children: [
-              TextFormField(
-                decoration: InputDecoration(
-                  contentPadding: const EdgeInsets.all(10),
-                  errorText: error.value == '' ? null : error.value,
-                  errorStyle: const TextStyle(fontSize: 14),
-                  labelText: locales.password,
-                  border: const OutlineInputBorder(),
-                  suffix: InkWell(
-                    onTap: () {
-                      if (obscure.value) {
-                        obscure.value = false;
-                      } else {
-                        obscure.value = true;
-                      }
-                    },
-                    child: Padding(
-                      padding: const EdgeInsets.only(right: 10.0),
-                      child: Icon(
-                        obscure.value ? Icons.visibility_off : Icons.visibility,
-                      ),
+        Column(
+          children: [
+            TextFormField(
+              decoration: InputDecoration(
+                contentPadding: const EdgeInsets.all(10),
+                errorText: error.value,
+                errorStyle: const TextStyle(fontSize: 14),
+                labelText: context.loc.password,
+                border: const OutlineInputBorder(),
+                suffix: Padding(
+                  padding: const EdgeInsets.all(5),
+                  child: SizedBox(
+                    height: 20,
+                    width: 20,
+                    child: IconButton(
+                      iconSize: 25,
+                      // ignore: use_named_constants
+                      padding: const EdgeInsets.all(0),
+                      color: context.colorScheme.primary,
+                      icon: obscure.value
+                          ? const Icon(
+                              Icons.visibility,
+                            )
+                          : const Icon(
+                              Icons.visibility_off,
+                            ),
+                      onPressed: () => obscure.value = !obscure.value,
                     ),
                   ),
                 ),
-                controller: passController,
-                obscureText: obscure.value,
               ),
-              const Padding(padding: EdgeInsets.only(bottom: 10)),
-              TextFormField(
-                decoration: InputDecoration(
-                  contentPadding: const EdgeInsets.all(10),
-                  errorStyle: const TextStyle(fontSize: 14),
-                  labelText: locales.confirmPassword,
-                  border: const OutlineInputBorder(),
-                  suffix: InkWell(
-                    onTap: () {
-                      if (obscure.value) {
-                        obscure.value = false;
-                      } else {
-                        obscure.value = true;
-                      }
-                    },
-                    child: Padding(
-                      padding: const EdgeInsets.only(right: 10.0),
-                      child: Icon(
-                        obscure.value ? Icons.visibility_off : Icons.visibility,
-                      ),
+              controller: passController,
+              obscureText: obscure.value,
+            ),
+            const Padding(padding: EdgeInsets.only(bottom: 10)),
+            TextFormField(
+              decoration: InputDecoration(
+                contentPadding: const EdgeInsets.all(10),
+                errorStyle: const TextStyle(fontSize: 14),
+                labelText: context.loc.confirmPassword,
+                border: const OutlineInputBorder(),
+                suffix: Padding(
+                  padding: const EdgeInsets.all(5),
+                  child: SizedBox(
+                    height: 20,
+                    width: 20,
+                    child: IconButton(
+                      iconSize: 25,
+                      // ignore: use_named_constants
+                      padding: const EdgeInsets.all(0),
+                      color: context.colorScheme.primary,
+                      icon: obscure.value
+                          ? const Icon(
+                              Icons.visibility,
+                            )
+                          : const Icon(
+                              Icons.visibility_off,
+                            ),
+                      onPressed: () => obscure.value = !obscure.value,
                     ),
                   ),
                 ),
-                controller: confirmPassController,
-                obscureText: obscure.value,
               ),
-              const Padding(padding: EdgeInsets.only(bottom: 10)),
-              Text(
-                locales.passwordCriteria,
-                style: const TextStyle(color: Colors.grey),
-              )
-            ],
-          ),
+              controller: confirmPassController,
+              obscureText: obscure.value,
+            ),
+            const Padding(padding: EdgeInsets.only(bottom: 10)),
+            Text(
+              context.loc.passwordCriteria,
+              style: const TextStyle(color: Colors.grey),
+            )
+          ],
         ),
       ],
-      onButtonPress: () async {
-        await Core.auth.signUp(
-            email: email,
-            password: passController.text,
-            confirmPassword: confirmPassController.text,
-            displayName: displayName,
-            username: username,
-            error: error,
-            context: context);
+      action: () {
+        final state = ref.read(signupState);
+        Core.auth.signUp(
+          email: state.email!,
+          password: passController.text,
+          confirmPassword: confirmPassController.text,
+          displayName: state.name!,
+          username: state.username!,
+          error: error,
+          context: context,
+        );
       },
-      isAsync: true,
     );
   }
 }
