@@ -1,69 +1,67 @@
-import 'package:allo/components/oobe_page.dart';
-import 'package:allo/generated/l10n.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class SetupUsername extends HookWidget {
-  const SetupUsername(
-      {required this.displayName, required this.email, Key? key})
-      : super(key: key);
-  final String displayName;
-  final String email;
+final _usernameReg = RegExp(r'^[a-zA-Z0-9_\.]+$');
+
+class SetupUsername extends HookConsumerWidget {
+  const SetupUsername({
+    super.key,
+  });
 
   @override
-  Widget build(BuildContext context) {
-    final locales = S.of(context);
+  Widget build(context, ref) {
     final controller = useTextEditingController();
-    final error = useState('');
+    final focusNode = useFocusNode();
+    final error = useState<String?>(null);
+
+    void onSubmit() async {
+      Core.auth.isUsernameCompliant(
+        username: controller.text,
+        error: error,
+        context: context,
+        focusNode: focusNode,
+        ref: ref,
+      );
+    }
+
     return SetupPage(
-      header: [
-        Text(
-          locales.setupUsernameScreenTitle,
-          style: const TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-          textAlign: TextAlign.left,
-        ),
-        const Padding(
-          padding: EdgeInsets.only(top: 10),
-        ),
-        Text(
-          locales.setupUsernameScreenDescription,
-          style: const TextStyle(fontSize: 17, color: Colors.grey),
-          textAlign: TextAlign.left,
-        ),
-      ],
+      icon: Icons.person_search,
+      title: Text(context.loc.setupUsernameScreenTitle),
+      subtitle: Text(context.loc.setupUsernameScreenDescription),
       body: [
-        Padding(
-          padding: const EdgeInsets.only(left: 20, right: 20),
-          child: Column(
-            children: [
-              TextFormField(
-                decoration: InputDecoration(
-                  contentPadding: const EdgeInsets.all(10),
-                  prefix: const Text('@'),
-                  errorText: error.value == '' ? null : error.value,
-                  errorStyle: const TextStyle(fontSize: 14),
-                  labelText: locales.username,
-                  border: const OutlineInputBorder(),
-                ),
-                controller: controller,
-              ),
-              const Padding(padding: EdgeInsets.only(bottom: 10)),
-              Text(
-                locales.setupUsernameRequirements,
-                style: const TextStyle(color: Colors.grey),
-              )
-            ],
+        TextFormField(
+          focusNode: focusNode,
+          autofocus: true,
+          decoration: InputDecoration(
+            contentPadding: const EdgeInsets.all(10),
+            prefix: const Text('@'),
+            errorText: error.value,
+            errorStyle: const TextStyle(fontSize: 14),
+            labelText: context.loc.username,
+            border: const OutlineInputBorder(),
           ),
+          onChanged: (_) {
+            if (!_usernameReg.hasMatch(_)) {
+              error.value = context.loc.errorThisIsInvalid(
+                context.loc.username.toLowerCase(),
+              );
+            } else {
+              error.value = null;
+            }
+          },
+          onFieldSubmitted: (_) => onSubmit(),
+          controller: controller,
+        ),
+        const Padding(padding: EdgeInsets.only(bottom: 10)),
+        Text(
+          context.loc.setupUsernameRequirements,
+          style: const TextStyle(color: Colors.grey),
         )
       ],
-      onButtonPress: () async => Core.auth.isUsernameCompliant(
-          username: controller.text,
-          error: error,
-          context: context,
-          displayName: displayName,
-          email: email),
-      isAsync: true,
+      action: onSubmit,
     );
   }
 }
